Fikse spinner
- Fikset spinner i POC 1
- Fikset spinner i POC 2
- Sarafiks i CSS
- Fikset spinner POC 3
- Ryddet litt opp i kode
- Sara fikset å splitte script.js filen i flere filer
