+++ inspill-poc2/frontend/oppsummering.html
@@ -17,7 +17,7 @@
 
         <div class="content-wrapper">
             <!-- Dokumenter container -->
-            <div class="content-area3">
+            <div class="content-area3 content-area3--small">
                 <div id="content">
                     <div class="containers-grid">
                     </div>
@@ -31,7 +31,7 @@
             </div>
             
             <!-- Oppsumerings container -->
-            <div class="content-area4">
+            <div class="content-area content-area3--medium">
                 <div id="content">
                     <div class="containers-grid">
                     </div>
@@ -44,7 +44,7 @@
             </div>
             
             <!-- Kategori container -->
-            <div class="content-area3-">
+            <div class="content-area3 content-area3--small">
                 <div id="content">
                     <div class="containers-grid">
                     </div>
