+++ sjekkliste-poc3/backend/app.py
@@ -8,14 +8,17 @@ from services.analysis_service import extract_checklist_points, load_checklist_f
 from read_pdf import read_pdf
 app = Flask(__name__)
 CORS(app)
+
 # Mappe for opplastede filer
 UPLOAD_FOLDER = Path(__file__).parent / "uploads"
 UPLOAD_FOLDER.mkdir(exist_ok=True)
 app.config["UPLOAD_FOLDER"] = str(UPLOAD_FOLDER)
+
 # Hold oversikt over den siste opplastingen i denne kjøretiden
 LAST_UPLOADS = []
 SJEKKLISTER_DIR = Path(__file__).parent / "sjekklister"
 DEFAULT_CHECKLIST = "sjekkliste_for_planbeskrivelse_bokm_mal.pdf"
+
 def _resolve_checklist_filename(requested_name: str | None) -> str:
     if not requested_name:
         return DEFAULT_CHECKLIST
@@ -24,10 +27,12 @@ def _resolve_checklist_filename(requested_name: str | None) -> str:
     if path.is_file():
         return safe_name
     raise FileNotFoundError(f"Sjekkliste finnes ikke: {safe_name}")
+
 def _clear_uploads(folder: Path) -> None:
     for item in folder.iterdir():
         if item.is_file():
             item.unlink(missing_ok=True)
+
 #tar imot én fil fra frontend, sjekker om den finnes fra før, og lagrer den hvis den er ny
 @app.route("/upload", methods=["POST"])
 def upload():
@@ -42,17 +47,20 @@ def upload():
         filename = secure_filename(file.filename)
         path = os.path.join(app.config["UPLOAD_FOLDER"], filename)
         file.save(path)
+        
         # Oppdater siste opplastning
         global LAST_UPLOADS
         LAST_UPLOADS = [filename]
         return jsonify({'uploaded': [filename]}), 200
     except Exception as e:
         return jsonify({'error': str(e)}), 500
+
 #returnerer en liste over alle filer i uploads-mappen
 @app.route("/documents", methods=["GET"])
 def get_documents():
     docs = [f.name for f in UPLOAD_FOLDER.iterdir() if f.is_file()]
     return jsonify(docs), 200
+
 #Server en spesifikk fil fra uploads-mappen.
 @app.route("/uploads/<filename>", methods=["GET"])
 def serve_file(filename):
@@ -60,6 +68,7 @@ def serve_file(filename):
         return send_from_directory(app.config["UPLOAD_FOLDER"], filename)
     except Exception:
         return jsonify({"error": "Filen finnes ikke"}), 404
+
 @app.route("/checklist", methods=["GET"])
 def get_checklist():
     try:
@@ -72,6 +81,7 @@ def get_checklist():
         return jsonify({"error": str(e)}), 400
     except Exception as e:
         return jsonify({"error": str(e)}), 500
+
 #returner KI analyse
 @app.route("/analysis", methods=["GET"])
 def get_analysis():
