+++ sjekkliste-poc3/frontend/index.html
@@ -31,7 +31,9 @@
     <button type="submit" class="analyse-btn">Sjekk planbeskrivelse</button>
 </form>
 </div>
-<script src="js/script.js"></script>
+<script type="module" src="js/script.js"></script>
+<script type="module" src="js/index.js"></script>
+
 </body>
 
 </html>
