+++ sjekkliste-poc3/frontend/sjekkliste.html
@@ -37,7 +37,8 @@
             </div>
         </div>
     </div>
-    <script src="js/script.js"></script>
+    <script type="module" src="js/script.js"></script>
+    <script type="module" src="js/sjekkliste.js"></script>
 
 </body>
